Fixed Linux compilation problems

+++ Source/Assets Compiler/Fonts/Font.cs
@@ -126,7 +126,7 @@ namespace Pegasus.AssetsCompiler.Fonts
 
 			var kerning = new FreeType.Vector();
 			FreeType.Invoke(() => FreeType.GetKerning(_fontPtr, left.Index, right.Index, 0, out kerning));
-			return kerning.x / 64;
+			return (int)kerning.x / 64;
 		}
 
 		/// <summary>
